Remove "winner" from results report again
You can add this back in your own versions if you want, but it's a very
misleading statistic that has nothing to do with your success in the
tournament. You can end up with a very low score and end up "winning"
most or all of the individual rounds, or end up with a high score and
end up "winning" close to none. It only measures whether your score is
higher than your opponent's, but not how high it actually is.